separated out a method that uses logout to make it run last [MQ-78][MQ-79][MQ-80][MQ-81][MQ-82]
